// src/data/biblia-nvi.type.ts
export type Versiculo = string;

export type Capitulo = Versiculo[];

export type NomeLivro = 
  | 'Gênesis' 
  | 'Êxodo' 
  | 'Levítico' 
  | 'Números' 
  | 'Deuteronômio' 
  | 'Josué' 
  | 'Juízes' 
  | 'Rute' 
  | '1 Samuel' 
  | '2 Samuel' 
  | '1 Reis' 
  | '2 Reis' 
  | '1 Crônicas' 
  | '2 Crônicas' 
  | 'Esdras' 
  | 'Neemias' 
  | 'Ester' 
  | 'Jó' 
  | 'Salmos' 
  | 'Provérbios' 
  | 'Eclesiastes' 
  | 'Cânticos' 
  | 'Isaías' 
  | 'Jeremias' 
  | 'Lamentações' 
  | 'Ezequiel' 
  | 'Daniel' 
  | 'Oséias' 
  | 'Joel' 
  | 'Amós' 
  | 'Obadias' 
  | 'Jonas' 
  | 'Miquéias' 
  | 'Naum' 
  | 'Habacuque' 
  | 'Sofonias' 
  | 'Ageu' 
  | 'Zacarias' 
  | 'Malaquias'
  | 'Mateus'
  | 'Marcos'
  | 'Lucas'
  | 'João'
  | 'Atos'
  | 'Romanos'
  | '1 Coríntios'
  | '2 Coríntios'
  | 'Gálatas'
  | 'Efésios'
  | 'Filipenses'
  | 'Colossenses'
  | '1 Tessalonicenses'
  | '2 Tessalonicenses'
  | '1 Timóteo'
  | '2 Timóteo'
  | 'Tito'
  | 'Filemom'
  | 'Hebreus'
  | 'Tiago'
  | '1 Pedro'
  | '2 Pedro'
  | '1 João'
  | '2 João'
  | '3 João'
  | 'Judas'
  | 'Apocalipse';

  export type AbreviacaoLivro = 
  | 'gn'
  | 'ex'
  | 'lv'
  | 'nm'
  | 'dt'
  | 'js'
  | 'jz'
  | 'rt'
  | '1sm'
  | '2sm'
  | '1rs'
  | '2rs'
  | '1cr'
  | '2cr'
  | 'ed'
  | 'ne'
  | 'et'
  | 'jó'
  | 'sl'
  | 'pv'
  | 'ec'
  | 'ct'
  | 'is'
  | 'jr'
  | 'lm'
  | 'ez'
  | 'dn'
  | 'os'
  | 'jl'
  | 'am'
  | 'ob'
  | 'jn'
  | 'mq'
  | 'na'
  | 'hc'
  | 'sf'
  | 'ag'
  | 'zc'
  | 'ml'
  | 'mt'
  | 'mc'
  | 'lc'
  | 'jo'
  | 'atos'
  | 'rm'
  | '1co'
  | '2co'
  | 'gl'
  | 'ef'
  | 'fp'
  | 'cl'
  | '1ts'
  | '2ts'
  | '1tm'
  | '2tm'
  | 'tt'
  | 'fm'
  | 'hb'
  | 'tg'
  | '1pe'
  | '2pe'
  | '1jo'
  | '2jo'
  | '3jo'
  | 'jd'
  | 'ap';

export interface LivroBiblico {
  name: NomeLivro;
  abbrev: AbreviacaoLivro;
  chapters: Capitulo[];
}


export type BibliaNVI = LivroBiblico[];